+++ server/data-import/data.js
@@ -5,11 +5,11 @@ const Log = require('log')
 const mongoose = require('mongoose')
 const P = require('bluebird')
 const fs = require('fs')
-const PregnancyCenterModel = require('../app/models/pregnancy-center')
+const PregnancyCenterModel = require('../pregnancy-centers/schema/mongoose-schema')
 const EJSON = require('mongodb-extended-json')
 const Joi = require('joi')
 P.promisifyAll(Joi)
-const pregnancyCenterSchemaJoi = require('../app/schemas/pregnancy-center')
+const pregnancyCenterSchemaJoi = require('../pregnancy-centers/schema/joi-schema')
 const log = new Log('info')
 mongoose.Promise = require('bluebird')
 P.promisifyAll(mongoose)
